refactor(policy): declare input image format on policies
