+++ src/lerobot/policies/__init__.py
@@ -28,7 +28,7 @@ from .multi_task_dit.configuration_multi_task_dit import MultiTaskDiTConfig as M
 from .pi0.configuration_pi0 import PI0Config as PI0Config
 from .pi0_fast.configuration_pi0_fast import PI0FastConfig as PI0FastConfig
 from .pi05.configuration_pi05 import PI05Config as PI05Config
-from .pretrained import PreTrainedPolicy as PreTrainedPolicy
+from .pretrained import ImageInputFormat as ImageInputFormat, PreTrainedPolicy as PreTrainedPolicy
 from .smolvla.configuration_smolvla import SmolVLAConfig as SmolVLAConfig
 from .tdmpc.configuration_tdmpc import TDMPCConfig as TDMPCConfig
 from .utils import make_robot_action, prepare_observation_for_inference
@@ -61,6 +61,7 @@ __all__ = [
     "WallXConfig",
     "XVLAConfig",
     # Base class
+    "ImageInputFormat",
     "PreTrainedPolicy",
     # RTC utilities
     "ActionInterpolator",
