+++ src/lerobot/processor/__init__.py
@@ -70,7 +70,6 @@ from .pipeline import (
     DataProcessorPipeline,
     DoneProcessorStep,
     IdentityProcessorStep,
-    ImageInputFormat,
     InfoProcessorStep,
     ObservationProcessorStep,
     PolicyActionProcessorStep,
@@ -126,7 +125,6 @@ __all__ = [
     "GymHILAdapterProcessorStep",
     "GripperPenaltyProcessorStep",
     "hotswap_stats",
-    "ImageInputFormat",
     "IdentityProcessorStep",
     "ImageCropResizeProcessorStep",
     "InfoProcessorStep",
