+++ tests/processor/test_pipeline.py
@@ -33,7 +33,6 @@ from lerobot.datasets.pipeline_features import aggregate_pipeline_dataset_featur
 from lerobot.processor import (
     DataProcessorPipeline,
     EnvTransition,
-    ImageInputFormat,
     ProcessorStep,
     ProcessorStepRegistry,
     TransitionKey,
@@ -605,18 +604,11 @@ def test_save_and_load_pretrained():
         assert loaded_pipeline.steps[1].counter == 10
 
 
-def test_input_image_format_is_backward_compatible_and_serialized():
-    legacy_pipeline = DataProcessorPipeline.from_config({"steps": []})
-    assert legacy_pipeline.input_image_format is ImageInputFormat.FLOAT32_0_1
+def test_policy_image_format_is_not_owned_by_pipeline():
+    pipeline = DataProcessorPipeline([])
 
-    pipeline = DataProcessorPipeline(
-        [],
-        input_image_format=ImageInputFormat.UINT8_0_255,
-    )
-    loaded_pipeline = DataProcessorPipeline.from_config(pipeline.get_config())
-
-    assert pipeline.get_config()["input_image_format"] == "uint8_0_255"
-    assert loaded_pipeline.input_image_format is ImageInputFormat.UINT8_0_255
+    assert not hasattr(pipeline, "input_image_format")
+    assert "input_image_format" not in pipeline.get_config()
 
 
 def test_step_without_optional_methods():
